+++ packages/exo-npm/index.js
@@ -15,4 +15,4 @@ export { makeMemoryCasStore, makeRetentionLinkSet } from './src/store.js';
 
 export { sha256HexWebCrypto } from './src/store-web-powers.js';
 
-export { makeJsReferenceRegistry } from './src/reference-backend.js';
+export { makeNpmReferenceRegistry } from './src/reference-backend.js';